+++ src/components/Floor.jsx
@@ -1,6 +1,7 @@
 import React from "react";
 import { hot } from "react-hot-loader";
 import PropTypes from "prop-types";
+
 import "./Floor.css";
 
 function Floor({ isFollowing, name }) {
