+++ src/index.js
@@ -1,8 +1,10 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import App from "./App";
-import store from "./store";
 import { Provider } from "react-redux";
+
+import store from "./redux/store";
+
+import App from "./components/App";
 
 ReactDOM.render(
   <Provider store={store}>
