+++ src/redux/store.js
@@ -1,4 +1,5 @@
 import { configureStore } from "@reduxjs/toolkit";
+
 import mainReducer from "./mainSlice";
 
 export default configureStore({
